refactor: Split up utils into utils and tokenization_utils
